Working on plotting.
